David dev: refactor classes to fit db schema

+++ TeamProject/src/main/java/com/restinginbed/TeamProject/jpa_repositories/OrganizationRepository.java
@@ -3,5 +3,5 @@ package com.restinginbed.TeamProject.jpa_repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.restinginbed.TeamProject.Organization;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Long> {
+public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
 }
